requirements.txt
numpy==1.16.2
scipy==1.2.3
sklearn

nltk==3.4
fuzzywuzzy[speedup]

rdflib
iribaker
SPARQLWrapper

playsound
webrtcvad
google-cloud-speech
google-cloud-TextToSpeech
google-cloud-translate

requests
pycountry
reverse_geocoder

pillow
opencv-python==4.2.0.32

tornado==5.1.1
typing
pyaudio

mock